test: consolidate and enhance core component test coverage
- Consolidate chunking tests into test_chunking_advanced.py with comprehensive edge cases
- Enhance client_manager.py coverage with advanced service creation tests
- Improve health_checks utilities test coverage for all service types
- Remove duplicate test files to streamline test suite organization
- Achieve ≥90% coverage target for core infrastructure components

This consolidation improves maintainability while ensuring comprehensive
test coverage for critical system components including chunking system,
client management, and health monitoring utilities.
