--- tests/unit/infrastructure/test_client_manager_advanced_coverage.py
@@ -1,474 +1,0 @@
-"""Advanced tests to boost client_manager.py coverage to 90%+.
-
-This test suite targets specific uncovered code paths in the client manager
-to achieve the ≥90% coverage goal.
-"""
-
-import pytest
-import asyncio
-from unittest.mock import Mock, patch, AsyncMock, MagicMock
-from src.infrastructure.client_manager import ClientManager
-from src.config.models import UnifiedConfig
-
-
-class TestABTestingManagerCreation:
-    """Test AB testing manager creation (lines 518-535)."""
-    
-    @pytest.mark.asyncio
-    async def test_get_ab_testing_manager_creation(self):
-        """Test creation of AB testing manager."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # Mock the ABTestingManager and its dependencies
-        with patch('src.services.deployment.ab_testing.ABTestingManager') as mock_ab_class, \
-             patch.object(client_manager, 'get_qdrant_service') as mock_get_qdrant, \
-             patch.object(client_manager, 'get_cache_manager') as mock_get_cache:
-            
-            mock_ab_instance = Mock()
-            mock_ab_class.return_value = mock_ab_instance
-            
-            mock_qdrant_service = Mock()
-            mock_cache_manager = Mock()
-            mock_get_qdrant.return_value = mock_qdrant_service
-            mock_get_cache.return_value = mock_cache_manager
-            
-            # First call should create the manager
-            ab_manager = await client_manager.get_ab_testing_manager()
-            
-            assert ab_manager is mock_ab_instance
-            mock_ab_class.assert_called_once_with(
-                qdrant_service=mock_qdrant_service,
-                cache_manager=mock_cache_manager
-            )
-            
-            # Second call should return the same instance (cached)
-            ab_manager2 = await client_manager.get_ab_testing_manager()
-            assert ab_manager2 is mock_ab_instance
-            
-            # Should not create a new instance
-            mock_ab_class.assert_called_once()
-    
-    @pytest.mark.asyncio
-    async def test_ab_testing_manager_concurrent_creation(self):
-        """Test concurrent creation of AB testing manager."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        with patch('src.services.deployment.ab_testing.ABTestingManager') as mock_ab_class, \
-             patch.object(client_manager, 'get_qdrant_service') as mock_get_qdrant, \
-             patch.object(client_manager, 'get_cache_manager') as mock_get_cache:
-            
-            mock_ab_instance = Mock()
-            mock_ab_class.return_value = mock_ab_instance
-            
-            mock_qdrant_service = Mock()
-            mock_cache_manager = Mock()
-            mock_get_qdrant.return_value = mock_qdrant_service
-            mock_get_cache.return_value = mock_cache_manager
-            
-            # Create multiple concurrent requests
-            tasks = [
-                client_manager.get_ab_testing_manager(),
-                client_manager.get_ab_testing_manager(),
-                client_manager.get_ab_testing_manager()
-            ]
-            
-            results = await asyncio.gather(*tasks)
-            
-            # All should return the same instance
-            assert all(result is mock_ab_instance for result in results)
-            
-            # Should only create one instance despite concurrent requests
-            mock_ab_class.assert_called_once()
-
-
-class TestQdrantClientCreation:
-    """Test Qdrant client creation (lines 703-712)."""
-    
-    @pytest.mark.asyncio
-    async def test_create_qdrant_client(self):
-        """Test creation of Qdrant client with configuration."""
-        config = UnifiedConfig()
-        config.qdrant.url = "http://localhost:6333"
-        config.qdrant.api_key = "test-key"
-        config.qdrant.timeout = 30.0
-        config.qdrant.prefer_grpc = True
-        
-        client_manager = ClientManager(config)
-        
-        with patch('src.infrastructure.client_manager.AsyncQdrantClient') as mock_qdrant_class:
-            mock_client = AsyncMock()
-            mock_client.get_collections = AsyncMock()
-            mock_qdrant_class.return_value = mock_client
-            
-            # Call the private method directly
-            qdrant_client = await client_manager._create_qdrant_client()
-            
-            assert qdrant_client is mock_client
-            mock_qdrant_class.assert_called_once_with(
-                url="http://localhost:6333",
-                api_key="test-key",
-                timeout=30.0,
-                prefer_grpc=True
-            )
-            
-            # Verify the validation call was made
-            mock_client.get_collections.assert_called_once()
-    
-    @pytest.mark.asyncio
-    async def test_qdrant_client_with_default_config(self):
-        """Test Qdrant client creation with default configuration."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        with patch('src.infrastructure.client_manager.AsyncQdrantClient') as mock_qdrant_class:
-            mock_client = AsyncMock()
-            mock_client.get_collections = AsyncMock()
-            mock_qdrant_class.return_value = mock_client
-            
-            await client_manager._create_qdrant_client()
-            
-            # Should be called with default config values
-            mock_qdrant_class.assert_called_once()
-            call_kwargs = mock_qdrant_class.call_args.kwargs
-            assert "url" in call_kwargs
-            assert "timeout" in call_kwargs
-            
-            # Verify validation was called
-            mock_client.get_collections.assert_called_once()
-
-
-class TestHealthCheckParallelExecution:
-    """Test parallel health check execution (lines 901-920)."""
-    
-    @pytest.mark.asyncio
-    async def test_parallel_health_checks_execution(self):
-        """Test parallel execution of health checks."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # Add some mock clients to trigger health checks
-        client_manager._clients["qdrant"] = Mock()
-        client_manager._clients["openai"] = Mock()
-        
-        # Mock the specific health check methods
-        async def mock_check_qdrant_health():
-            await asyncio.sleep(0.01)  # Simulate async operation
-            return True
-        
-        async def mock_check_openai_health():
-            await asyncio.sleep(0.01)  # Simulate async operation
-            return True
-        
-        with patch.object(client_manager, '_check_qdrant_health', side_effect=mock_check_qdrant_health), \
-             patch.object(client_manager, '_check_openai_health', side_effect=mock_check_openai_health), \
-             patch('asyncio.create_task', wraps=asyncio.create_task) as mock_create_task, \
-             patch('asyncio.gather', wraps=asyncio.gather) as mock_gather:
-            
-            # Mock the health check loop internal logic (lines 901-920)
-            health_checks = {
-                "qdrant": client_manager._check_qdrant_health,
-                "openai": client_manager._check_openai_health,
-            }
-            
-            # Simulate the parallel health check execution from lines 901-920
-            active_clients = [
-                (name, check_func)
-                for name, check_func in health_checks.items()
-                if name in client_manager._clients
-            ]
-            
-            # Create tasks for parallel execution (this is what the real code does)
-            tasks = []
-            for name, check_func in active_clients:
-                task = asyncio.create_task(
-                    client_manager._run_single_health_check(name, check_func),
-                    name=f"health_check_{name}",
-                )
-                tasks.append(task)
-            
-            # Wait for all health checks to complete
-            await asyncio.gather(*tasks, return_exceptions=True)
-            
-            # Should have created tasks for parallel execution
-            assert mock_create_task.call_count >= 2
-            assert mock_gather.called
-    
-    @pytest.mark.asyncio
-    async def test_health_checks_with_no_active_clients(self):
-        """Test health check behavior when no active clients exist."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # No clients added to _clients dictionary
-        assert len(client_manager._clients) == 0
-        
-        # Test the logic from lines 907-908 where no active clients exist
-        health_checks = {
-            "qdrant": client_manager._check_qdrant_health,
-            "openai": client_manager._check_openai_health,
-        }
-        
-        # Filter to active clients (should be empty)
-        active_clients = [
-            (name, check_func)
-            for name, check_func in health_checks.items()
-            if name in client_manager._clients
-        ]
-        
-        # Should be empty since no clients are registered
-        assert len(active_clients) == 0
-        
-        # This simulates the "continue" logic when no active clients exist
-        if not active_clients:
-            # This path should be taken (lines 907-908)
-            pass  # Test passes by reaching this point
-
-
-class TestServiceLockHandling:
-    """Test service lock handling and concurrent access."""
-    
-    @pytest.mark.asyncio
-    async def test_service_lock_creation_and_usage(self):
-        """Test creation and usage of service locks."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # Test lock creation for new service
-        service_name = "test_service"
-        
-        # Initially no lock should exist
-        assert service_name not in client_manager._service_locks
-        
-        # Mock a service creation method that uses locks
-        async def mock_service_creation():
-            if service_name not in client_manager._service_locks:
-                client_manager._service_locks[service_name] = asyncio.Lock()
-            
-            async with client_manager._service_locks[service_name]:
-                # Simulate service creation
-                await asyncio.sleep(0.01)
-                return "service_instance"
-        
-        # Test concurrent access
-        tasks = [
-            mock_service_creation(),
-            mock_service_creation(),
-            mock_service_creation()
-        ]
-        
-        results = await asyncio.gather(*tasks)
-        
-        # All should complete successfully
-        assert all(result == "service_instance" for result in results)
-        
-        # Lock should have been created
-        assert service_name in client_manager._service_locks
-        assert isinstance(client_manager._service_locks[service_name], asyncio.Lock)
-
-
-class TestErrorHandlingAndEdgeCases:
-    """Test error handling and edge cases in client manager."""
-    
-    @pytest.mark.asyncio
-    async def test_client_creation_with_circuit_breaker_errors(self):
-        """Test client creation when circuit breaker is in error state."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # Mock a circuit breaker in error state
-        mock_circuit_breaker = Mock()
-        mock_circuit_breaker.state = "open"  # Circuit breaker is open (error state)
-        mock_circuit_breaker.failure_count = 10
-        
-        client_manager._circuit_breakers["test_service"] = mock_circuit_breaker
-        
-        # Mock client creation function that would fail
-        async def failing_create_func():
-            raise Exception("Service unavailable")
-        
-        async def failing_health_check():
-            raise Exception("Health check failed")
-        
-        # This should handle circuit breaker logic
-        try:
-            await client_manager._get_or_create_client(
-                "test_service", 
-                failing_create_func, 
-                failing_health_check
-            )
-        except Exception:
-            # Expected to fail, but should handle circuit breaker state
-            pass
-        
-        # Circuit breaker should exist
-        assert "test_service" in client_manager._circuit_breakers
-    
-    @pytest.mark.asyncio
-    async def test_health_status_with_circuit_breaker_states(self):
-        """Test health status reporting with various circuit breaker states."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # Add clients with different circuit breaker states
-        client_manager._clients["healthy_service"] = Mock()
-        client_manager._clients["failing_service"] = Mock()
-        
-        # Mock circuit breakers with different states
-        from src.infrastructure.client_manager import ClientState, ClientHealth
-        import time
-        
-        # Add health entries for the clients
-        client_manager._health["healthy_service"] = ClientHealth(
-            state=ClientState.HEALTHY,
-            last_check=time.time(),
-            consecutive_failures=0
-        )
-        client_manager._health["failing_service"] = ClientHealth(
-            state=ClientState.FAILED,
-            last_check=time.time(),
-            last_error="Service unavailable",
-            consecutive_failures=5
-        )
-        
-        # Mock circuit breakers with different states
-        healthy_breaker = Mock()
-        healthy_breaker.state = ClientState.HEALTHY
-        
-        failing_breaker = Mock()
-        failing_breaker.state = ClientState.FAILED
-        
-        client_manager._circuit_breakers["healthy_service"] = healthy_breaker
-        client_manager._circuit_breakers["failing_service"] = failing_breaker
-        
-        # Get health status
-        status = await client_manager.get_health_status()
-        
-        # Should include health information for both services
-        assert "healthy_service" in status
-        assert "failing_service" in status
-        
-        # Check that the status includes the expected information
-        assert status["healthy_service"]["state"] == "healthy"
-        assert status["failing_service"]["state"] == "failed"
-        assert status["failing_service"]["consecutive_failures"] == 5
-
-
-class TestClientManagerConfiguration:
-    """Test client manager configuration handling."""
-    
-    def test_client_manager_initialization_with_custom_config(self):
-        """Test client manager initialization with custom configuration."""
-        config = UnifiedConfig()
-        # Use actual performance config fields
-        config.performance.max_concurrent_requests = 20
-        config.performance.request_timeout = 45.0
-        
-        client_manager = ClientManager(config)
-        
-        # Should store the configuration
-        assert client_manager.config is config
-        assert client_manager.config.performance.max_concurrent_requests == 20
-        assert client_manager.config.performance.request_timeout == 45.0
-        
-        # Should initialize internal structures
-        assert hasattr(client_manager, '_clients')
-        assert hasattr(client_manager, '_circuit_breakers')
-        assert hasattr(client_manager, '_service_locks')
-        assert isinstance(client_manager._clients, dict)
-        assert isinstance(client_manager._circuit_breakers, dict)
-        assert isinstance(client_manager._service_locks, dict)
-    
-    @pytest.mark.asyncio
-    async def test_cleanup_and_resource_management(self):
-        """Test cleanup and resource management."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # Add some mock clients
-        mock_client1 = Mock()
-        mock_client2 = Mock()
-        
-        client_manager._clients["service1"] = mock_client1
-        client_manager._clients["service2"] = mock_client2
-        
-        # Verify clients are stored
-        assert len(client_manager._clients) == 2
-        
-        # Test that clients can be accessed
-        assert "service1" in client_manager._clients
-        assert "service2" in client_manager._clients
-        
-        # Test cleanup (if such method exists)
-        if hasattr(client_manager, 'cleanup'):
-            await client_manager.cleanup()
-        
-        # Could test resource cleanup here if implemented
-
-
-class TestSpecificMissingLines:
-    """Test specific missing lines that weren't covered by other tests."""
-    
-    @pytest.mark.asyncio
-    async def test_circuit_breaker_initialization_edge_cases(self):
-        """Test circuit breaker initialization edge cases."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        async def test_create_func():
-            return "test_client"
-        
-        async def test_health_func():
-            return True
-        
-        # This should create circuit breaker with default settings (since the config uses getattr with defaults)
-        client = await client_manager._get_or_create_client(
-            "test_service", test_create_func, test_health_func
-        )
-        
-        assert client == "test_client"
-        assert "test_service" in client_manager._circuit_breakers
-        
-        # Check that the circuit breaker has the correct threshold (default is 5)
-        circuit_breaker = client_manager._circuit_breakers["test_service"]
-        assert hasattr(circuit_breaker, 'failure_threshold')
-        assert circuit_breaker.failure_threshold == 5  # Default value
-    
-    @pytest.mark.asyncio
-    async def test_service_manager_creation_patterns(self):
-        """Test various service manager creation patterns."""
-        config = UnifiedConfig()
-        client_manager = ClientManager(config)
-        
-        # Test cache manager creation (simplest - no dependencies)
-        with patch('src.services.cache.manager.CacheManager') as mock_cache_class:
-            mock_cache_instance = Mock()
-            mock_cache_instance.initialize = AsyncMock()
-            mock_cache_class.return_value = mock_cache_instance
-            
-            # First call should create the manager
-            cache_manager = await client_manager.get_cache_manager()
-            assert cache_manager is mock_cache_instance
-            
-            # Second call should return cached instance
-            cache_manager2 = await client_manager.get_cache_manager()
-            assert cache_manager2 is mock_cache_instance
-            
-            # Should only initialize once
-            mock_cache_instance.initialize.assert_called_once()
-        
-        # Test embedding manager creation (has client_manager dependency)
-        with patch('src.services.embeddings.manager.EmbeddingManager') as mock_embed_class:
-            mock_embed_instance = Mock()
-            mock_embed_instance.initialize = AsyncMock()
-            mock_embed_class.return_value = mock_embed_instance
-            
-            embedding_manager = await client_manager.get_embedding_manager()
-            assert embedding_manager is mock_embed_instance
-            
-            # Should pass both config and client_manager
-            mock_embed_class.assert_called_once_with(
-                config=config,
-                client_manager=client_manager,
-            )
