--- tests/unit/test_chunking_advanced_coverage.py
@@ -1,532 +1,0 @@
-"""Advanced tests to boost chunking.py coverage to 90%+.
-
-This test suite targets specific uncovered code paths in the chunking module
-to achieve the ≥90% coverage goal.
-"""
-
-import pytest
-from unittest.mock import Mock, patch, MagicMock
-from src.chunking import EnhancedChunker
-from src.config.models import ChunkingConfig
-from src.config.enums import ChunkingStrategy
-
-
-class TestTreeSitterImportFallbacks:
-    """Test tree-sitter import fallback scenarios."""
-    
-    def test_tree_sitter_unavailable_fallback(self):
-        """Test fallback when tree-sitter is completely unavailable."""
-        # Mock the import to fail at the top level
-        with patch.dict('sys.modules', {'tree_sitter': None}):
-            with patch('src.chunking.TREE_SITTER_AVAILABLE', False):
-                with patch('src.chunking.Parser', None):
-                    with patch('src.chunking.Node', None):
-                        # Create chunker - should work without tree-sitter
-                        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-                        chunker = EnhancedChunker(config)
-                        
-                        # AST chunking should fallback to basic chunking
-                        content = "def test_function():\n    return 'hello'"
-                        chunks = chunker.chunk_content(content, "Test", "http://test.com")
-                        
-                        assert len(chunks) > 0
-                        assert chunks[0]["content"]
-    
-    def test_python_parser_unavailable(self):
-        """Test when python parser is unavailable but tree-sitter is."""
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            with patch('src.chunking.PYTHON_AVAILABLE', False):
-                with patch('src.chunking.tspython', None):
-                    config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-                    chunker = EnhancedChunker(config)
-                    
-                    # Should detect language as python but fallback due to unavailable parser
-                    python_code = "def test_function():\n    print('hello')\n    return True"
-                    chunks = chunker.chunk_content(python_code, "test.py", "http://test.py")
-                    
-                    assert len(chunks) > 0
-    
-    def test_javascript_parser_unavailable(self):
-        """Test when javascript parser is unavailable."""
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            with patch('src.chunking.JAVASCRIPT_AVAILABLE', False):
-                with patch('src.chunking.tsjavascript', None):
-                    config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-                    chunker = EnhancedChunker(config)
-                    
-                    js_code = "function testFunction() { return 'hello'; }"
-                    chunks = chunker.chunk_content(js_code, "test.js", "http://test.js")
-                    
-                    assert len(chunks) > 0
-    
-    def test_typescript_parser_unavailable(self):
-        """Test when typescript parser is unavailable."""
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            with patch('src.chunking.TYPESCRIPT_AVAILABLE', False):
-                with patch('src.chunking.tstypescript', None):
-                    config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-                    chunker = EnhancedChunker(config)
-                    
-                    ts_code = "function testFunction(): string { return 'hello'; }"
-                    chunks = chunker.chunk_content(ts_code, "test.ts", "http://test.ts")
-                    
-                    assert len(chunks) > 0
-
-
-class TestRecursiveLargeMethodSplitting:
-    """Test recursive splitting of large methods within classes."""
-    
-    def test_recursive_method_splitting_python(self):
-        """Test recursive splitting of very large methods in Python classes."""
-        config = ChunkingConfig(
-            strategy=ChunkingStrategy.AST, 
-            chunk_size=200,
-            chunk_overlap=20,
-            max_function_chunk_size=150
-        )
-        chunker = EnhancedChunker(config)
-        
-        # Create a very large Python method that exceeds max_function_chunk_size
-        large_method = """
-class TestClass:
-    def very_large_method(self):
-        # This method is intentionally very large
-        result = []
-        for i in range(100):
-            if i % 2 == 0:
-                result.append(i * 2)
-            else:
-                result.append(i * 3)
-            # Add more logic to make it larger
-            temp_value = i ** 2
-            if temp_value > 50:
-                result.append(temp_value)
-        
-        # More processing
-        processed_result = []
-        for item in result:
-            if item > 10:
-                processed_result.append(item / 2)
-            else:
-                processed_result.append(item * 2)
-        
-        # Final processing step
-        final_result = []
-        for val in processed_result:
-            if val > 5:
-                final_result.append(val + 1)
-            else:
-                final_result.append(val - 1)
-        
-        return final_result
-"""
-        
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            with patch('src.chunking.PYTHON_AVAILABLE', True):
-                # Mock tree-sitter components
-                mock_parser = Mock()
-                mock_tree = Mock()
-                mock_root = Mock()
-                
-                # Create mock class node
-                mock_class_node = Mock()
-                mock_class_node.type = "class_definition"
-                mock_class_node.start_byte = 0
-                mock_class_node.end_byte = len(large_method)
-                mock_class_node.children = []
-                
-                # Create mock method node that's too large
-                mock_method_node = Mock()
-                mock_method_node.type = "function_definition"
-                mock_method_node.start_byte = 50
-                mock_method_node.end_byte = len(large_method) - 10
-                mock_class_node.children = [mock_method_node]
-                
-                mock_root.children = [mock_class_node]
-                mock_tree.root_node = mock_root
-                mock_parser.parse.return_value = mock_tree
-                
-                with patch('src.chunking.Parser', return_value=mock_parser):
-                    with patch('src.chunking.tspython') as mock_tspython:
-                        mock_tspython.language.return_value = Mock()
-                        
-                        chunks = chunker.chunk_content(large_method, "test.py", "http://test.py")
-                        
-                        # Should have created chunks with recursive splitting
-                        assert len(chunks) > 0
-                        
-                        # Check for parent_class metadata indicating recursive splitting
-                        has_parent_class_metadata = any(
-                            chunk.get("metadata", {}).get("parent_class", False) 
-                            for chunk in chunks
-                        )
-                        # Note: This might not always be True depending on exact splitting logic
-    
-    def test_method_splitting_with_metadata_preservation(self):
-        """Test that parent class metadata is preserved during recursive splitting."""
-        config = ChunkingConfig(
-            strategy=ChunkingStrategy.AST,
-            chunk_size=100,
-            chunk_overlap=10,
-            max_function_chunk_size=80
-        )
-        chunker = EnhancedChunker(config)
-        
-        # Create content that will trigger recursive splitting
-        content = """
-class ParentClass:
-    def large_method(self):
-        # This method should be split recursively
-        step1 = "first part of processing"
-        step2 = "second part of processing"  
-        step3 = "third part of processing"
-        step4 = "fourth part of processing"
-        step5 = "fifth part of processing"
-        return step1 + step2 + step3 + step4 + step5
-"""
-        
-        # Mock the _split_large_code_unit method to return chunks with metadata
-        def mock_split_large_unit(content, start_pos, unit_type, language):
-            from src.chunking import Chunk
-            chunk1 = Chunk(content=content[:50], start_pos=start_pos, end_pos=start_pos+50)
-            chunk2 = Chunk(content=content[50:], start_pos=start_pos+50, end_pos=start_pos+len(content))
-            return [chunk1, chunk2]
-        
-        with patch.object(chunker, '_split_large_code_unit', side_effect=mock_split_large_unit):
-            chunks = chunker.chunk_content(content, "test.py", "http://test.py")
-            
-            # Should have created multiple chunks
-            assert len(chunks) > 0
-
-
-class TestFunctionBlockExtraction:
-    """Test the extract_function_blocks method that handles lines 1115-1202."""
-    
-    def test_extract_function_blocks_python_with_body(self):
-        """Test function block extraction for Python functions with body."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST, chunk_size=500)
-        chunker = EnhancedChunker(config)
-        
-        # Mock tree-sitter components
-        mock_node = Mock()
-        mock_node.start_byte = 0
-        mock_node.end_byte = 200
-        
-        # Mock body node
-        mock_body = Mock()
-        mock_body.type = "block"
-        mock_body.start_byte = 50
-        mock_body.end_byte = 190
-        
-        # Mock child nodes representing statements
-        mock_if_stmt = Mock()
-        mock_if_stmt.type = "if_statement"
-        mock_if_stmt.start_byte = 60
-        mock_if_stmt.end_byte = 100
-        
-        mock_for_stmt = Mock()
-        mock_for_stmt.type = "for_statement"
-        mock_for_stmt.start_byte = 110
-        mock_for_stmt.end_byte = 150
-        
-        mock_body.children = [mock_if_stmt, mock_for_stmt]
-        mock_node.children = [mock_body]
-        
-        # Test the method
-        test_content = "def test_function():\n    if True:\n        pass\n    for i in range(10):\n        pass"
-        blocks = chunker._extract_function_blocks(mock_node, test_content, "python")
-        
-        # Should return blocks based on the structure
-        assert isinstance(blocks, list)
-        assert len(blocks) > 0
-    
-    def test_extract_function_blocks_javascript_with_body(self):
-        """Test function block extraction for JavaScript functions."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST, chunk_size=500)
-        chunker = EnhancedChunker(config)
-        
-        mock_node = Mock()
-        mock_node.start_byte = 0
-        mock_node.end_byte = 200
-        
-        # Mock statement_block for JavaScript
-        mock_body = Mock()
-        mock_body.type = "statement_block"
-        mock_body.start_byte = 50
-        mock_body.end_byte = 190
-        
-        # Mock child statements
-        mock_try_stmt = Mock()
-        mock_try_stmt.type = "try_statement"
-        mock_try_stmt.start_byte = 60
-        mock_try_stmt.end_byte = 120
-        
-        mock_switch_stmt = Mock()
-        mock_switch_stmt.type = "switch_statement"
-        mock_switch_stmt.start_byte = 130
-        mock_switch_stmt.end_byte = 180
-        
-        mock_body.children = [mock_try_stmt, mock_switch_stmt]
-        mock_node.children = [mock_body]
-        
-        test_content = "function test() {\n    try {\n        doSomething();\n    } catch (e) {\n        handleError(e);\n    }\n    switch (x) {\n        case 1: break;\n    }\n}"
-        blocks = chunker._extract_function_blocks(mock_node, test_content, "javascript")
-        
-        assert isinstance(blocks, list)
-        assert len(blocks) > 0
-    
-    def test_extract_function_blocks_no_body_found(self):
-        """Test when no function body is found."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        # Mock node without body
-        mock_node = Mock()
-        mock_node.start_byte = 0
-        mock_node.end_byte = 100
-        mock_node.children = []  # No body node
-        
-        test_content = "def empty_function(): pass"
-        blocks = chunker._extract_function_blocks(mock_node, test_content, "python")
-        
-        # Should return the whole function
-        assert len(blocks) == 1
-        assert blocks[0]["type"] == "whole_function"
-        assert blocks[0]["start_pos"] == 0
-        assert blocks[0]["end_pos"] == 100
-    
-    def test_extract_function_blocks_single_block_result(self):
-        """Test when extraction results in only one block."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST, chunk_size=1000)
-        chunker = EnhancedChunker(config)
-        
-        mock_node = Mock()
-        mock_node.start_byte = 0
-        mock_node.end_byte = 100
-        
-        mock_body = Mock()
-        mock_body.type = "block"
-        mock_body.start_byte = 20
-        mock_body.end_byte = 90
-        
-        # Small child that doesn't need splitting
-        mock_child = Mock()
-        mock_child.type = "expression_statement"
-        mock_child.start_byte = 25
-        mock_child.end_byte = 30
-        
-        mock_body.children = [mock_child]
-        mock_node.children = [mock_body]
-        
-        test_content = "def small_function():\n    return True"
-        blocks = chunker._extract_function_blocks(mock_node, test_content, "python")
-        
-        # Should return whole function when only one block
-        assert len(blocks) == 1
-        assert blocks[0]["type"] == "whole_function"
-    
-    def test_extract_function_blocks_with_signature_inclusion(self):
-        """Test that function signature is included with first block."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST, chunk_size=200, chunk_overlap=20)
-        chunker = EnhancedChunker(config)
-        
-        mock_node = Mock()
-        mock_node.start_byte = 0
-        mock_node.end_byte = 300
-        
-        mock_body = Mock()
-        mock_body.type = "block"
-        mock_body.start_byte = 50
-        mock_body.end_byte = 280
-        
-        # Create children that will result in multiple blocks
-        mock_child1 = Mock()
-        mock_child1.type = "if_statement"
-        mock_child1.start_byte = 60
-        mock_child1.end_byte = 150
-        
-        mock_child2 = Mock() 
-        mock_child2.type = "for_statement"
-        mock_child2.start_byte = 160
-        mock_child2.end_byte = 250
-        
-        mock_body.children = [mock_child1, mock_child2]
-        mock_node.children = [mock_body]
-        
-        test_content = "def complex_function():\n    if True:\n        pass\n    for i in range(10):\n        pass"
-        blocks = chunker._extract_function_blocks(mock_node, test_content, "python")
-        
-        # Should have multiple blocks
-        assert len(blocks) > 1
-        # First block should start at function start (includes signature)
-        assert blocks[0]["start_pos"] == 0
-
-
-class TestASTParsingSpeicialCases:
-    """Test special cases in AST parsing that aren't covered."""
-    
-    def test_ast_parsing_with_parser_errors(self):
-        """Test AST parsing when parser encounters errors."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        content = "invalid python syntax $$$ def broken():"
-        
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            with patch('src.chunking.PYTHON_AVAILABLE', True):
-                mock_parser = Mock()
-                # Make parser.parse raise an exception
-                mock_parser.parse.side_effect = Exception("Parser error")
-                
-                with patch('src.chunking.Parser', return_value=mock_parser):
-                    with patch('src.chunking.tspython') as mock_tspython:
-                        mock_tspython.language.return_value = Mock()
-                        
-                        # Should fallback to basic chunking
-                        chunks = chunker.chunk_content(content, "test.py", "http://test.py")
-                        
-                        assert len(chunks) > 0
-                        assert chunks[0]["content"]
-    
-    def test_extract_code_units_empty_traversal(self):
-        """Test when AST traversal yields no code units."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        # Create mock tree with empty traversal
-        mock_tree = Mock()
-        mock_root = Mock()
-        mock_root.children = []  # No children, empty traversal
-        mock_tree.root_node = mock_root
-        
-        # Mock walk method to return empty iterator
-        mock_tree.walk = Mock(return_value=iter([]))
-        
-        code_units = chunker._extract_code_units(mock_tree, "python", 1000)
-        
-        assert isinstance(code_units, list)
-        assert len(code_units) == 0
-
-
-class TestChunkingBoundaryEdgeCases:
-    """Test edge cases in boundary detection and chunking logic."""
-    
-    def test_find_enhanced_boundary_no_patterns(self):
-        """Test enhanced boundary detection when no patterns match."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.ENHANCED)
-        chunker = EnhancedChunker(config)
-        
-        # Content with no clear boundaries
-        content = "This is just plain text without any clear boundaries or patterns."
-        
-        boundary = chunker._find_enhanced_boundary(content, 30, 50, "unknown")
-        
-        # Should return a reasonable boundary
-        assert isinstance(boundary, int)
-        assert 30 <= boundary <= 50
-    
-    def test_chunk_content_with_very_long_single_line(self):
-        """Test chunking content that is one very long line."""
-        config = ChunkingConfig(
-            strategy=ChunkingStrategy.BASIC,
-            chunk_size=100,
-            chunk_overlap=10
-        )
-        chunker = EnhancedChunker(config)
-        
-        # Create a very long single line
-        very_long_line = "This is a very long line that exceeds the chunk size limit but has no natural break points or boundaries that would allow for clean splitting at word or sentence boundaries." * 10
-        
-        chunks = chunker.chunk_content(very_long_line, "test.txt", "http://test.txt")
-        
-        assert len(chunks) > 1  # Should be split despite being one line
-        assert all(chunk["content"] for chunk in chunks)  # All chunks should have content
-    
-    def test_chunking_with_mixed_content_types(self):
-        """Test chunking with mixed code and text content."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.ENHANCED, chunk_size=200)
-        chunker = EnhancedChunker(config)
-        
-        mixed_content = """
-This is some explanatory text.
-
-```python
-def example_function():
-    return "Hello World"
-```
-
-More text explaining the code above.
-
-```javascript
-function anotherExample() {
-    console.log("Hello from JS");
-}
-```
-
-Final explanatory text.
-"""
-        
-        chunks = chunker.chunk_content(mixed_content, "mixed.md", "http://mixed.md")
-        
-        assert len(chunks) > 0
-        # Should preserve code blocks during chunking
-        assert any("def example_function" in chunk["content"] for chunk in chunks)
-    
-    def test_ast_chunking_unsupported_language_fallback(self):
-        """Test AST chunking fallback for unsupported languages."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        # Content in a language not supported by AST parsing
-        rust_code = """
-fn main() {
-    println!("Hello, world!");
-}
-
-fn another_function() -> i32 {
-    42
-}
-"""
-        
-        chunks = chunker.chunk_content(rust_code, "test.rs", "http://test.rs")
-        
-        # Should fallback to enhanced chunking
-        assert len(chunks) > 0
-        assert chunks[0]["content"]
-
-
-class TestChunkingConfigValidation:
-    """Test chunking configuration validation and edge cases."""
-    
-    def test_chunk_size_boundary_validation(self):
-        """Test validation of chunk size boundaries."""
-        # Test maximum boundary case
-        config = ChunkingConfig(
-            chunk_size=8000,  # Near maximum
-            chunk_overlap=100
-        )
-        chunker = EnhancedChunker(config)
-        
-        large_content = "word " * 2000  # Create large content
-        chunks = chunker.chunk_content(large_content, "large.txt", "http://large.txt")
-        
-        assert len(chunks) > 0
-        # Verify chunks respect size limits
-        for chunk in chunks:
-            assert len(chunk["content"]) <= config.chunk_size + config.chunk_overlap
-    
-    def test_zero_chunk_overlap_handling(self):
-        """Test handling of zero chunk overlap."""
-        config = ChunkingConfig(
-            chunk_size=100,
-            chunk_overlap=0
-        )
-        chunker = EnhancedChunker(config)
-        
-        content = "This is a test content that should be split into multiple chunks without any overlap between them."
-        chunks = chunker.chunk_content(content, "test.txt", "http://test.txt")
-        
-        assert len(chunks) > 1
-        # With zero overlap, chunks should not share content
-        for i in range(len(chunks) - 1):
-            assert chunks[i]["content"] != chunks[i + 1]["content"]
