--- tests/unit/test_chunking_ast_comprehensive.py
@@ -1,682 +1,0 @@
-"""Comprehensive tests for AST-based chunking features."""
-
-import pytest
-from unittest.mock import Mock, patch, MagicMock
-
-from src.chunking import EnhancedChunker
-from src.config.enums import ChunkingStrategy
-from src.config.models import ChunkingConfig
-from src.models.document_processing import Chunk
-
-
-class TestTreeSitterImports:
-    """Test Tree-sitter import handling and fallback behavior."""
-    
-    def test_initialization_without_tree_sitter(self):
-        """Test chunker initialization when Tree-sitter is not available."""
-        config = ChunkingConfig(enable_ast_chunking=True)
-        
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', False):
-            chunker = EnhancedChunker(config)
-            assert chunker.parsers == {}
-    
-    def test_initialization_with_unavailable_parsers(self):
-        """Test initialization when specific language parsers are unavailable."""
-        config = ChunkingConfig(
-            enable_ast_chunking=True,
-            supported_languages=["python", "javascript", "typescript"]
-        )
-        
-        # Mock Tree-sitter as available but specific parsers as unavailable
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True), \
-             patch('src.chunking.PYTHON_AVAILABLE', False), \
-             patch('src.chunking.JAVASCRIPT_AVAILABLE', False), \
-             patch('src.chunking.TYPESCRIPT_AVAILABLE', False):
-            chunker = EnhancedChunker(config)
-            assert chunker.parsers == {}
-    
-    def test_initialization_with_unsupported_language(self):
-        """Test initialization with unsupported language in config."""
-        config = ChunkingConfig(
-            enable_ast_chunking=True,
-            supported_languages=["ruby", "go"]  # Unsupported languages
-        )
-        
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            chunker = EnhancedChunker(config)
-            assert chunker.parsers == {}
-    
-    def test_parser_initialization_failure(self):
-        """Test handling of parser initialization failure."""
-        config = ChunkingConfig(
-            enable_ast_chunking=True,
-            supported_languages=["python"]
-        )
-        
-        mock_parser_module = Mock()
-        mock_parser_module.language.side_effect = Exception("Parser failed")
-        
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True), \
-             patch('src.chunking.PYTHON_AVAILABLE', True), \
-             patch('src.chunking.tspython', mock_parser_module), \
-             patch('src.chunking.Language') as mock_language, \
-             patch('src.chunking.Parser') as mock_parser:
-            
-            mock_language.side_effect = Exception("Language creation failed")
-            chunker = EnhancedChunker(config)
-            assert "python" not in chunker.parsers
-
-
-class TestASTParsing:
-    """Test AST-based parsing and code unit extraction."""
-    
-    def test_ast_chunking_fallback_without_parsers(self):
-        """Test AST chunking falls back to enhanced chunking when no parsers available."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        chunker.parsers = {}  # No parsers available
-        
-        content = """
-def test_function():
-    return "hello"
-
-class TestClass:
-    def method(self):
-        pass
-"""
-        
-        with patch.object(chunker, '_enhanced_chunking') as mock_enhanced:
-            mock_enhanced.return_value = [
-                Chunk(content="mock chunk", start_pos=0, end_pos=50, chunk_index=0)
-            ]
-            
-            chunks = chunker._ast_based_chunking(content, "python")
-            mock_enhanced.assert_called_once_with(content, "python")
-    
-    def test_ast_chunking_with_mock_parser(self):
-        """Test AST chunking with mocked Tree-sitter parser."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST, chunk_size=1000, chunk_overlap=100)
-        chunker = EnhancedChunker(config)
-        
-        # Mock parser and tree structure
-        mock_parser = Mock()
-        mock_tree = Mock()
-        mock_root = Mock()
-        
-        # Mock function node
-        mock_func_node = Mock()
-        mock_func_node.type = "function_definition"
-        mock_func_node.start_byte = 0
-        mock_func_node.end_byte = 50
-        mock_func_node.children = [
-            Mock(type="identifier", start_byte=4, end_byte=17)  # function name
-        ]
-        
-        # Mock class node
-        mock_class_node = Mock()
-        mock_class_node.type = "class_definition"
-        mock_class_node.start_byte = 52
-        mock_class_node.end_byte = 100
-        mock_class_node.children = [
-            Mock(type="identifier", start_byte=58, end_byte=67)  # class name
-        ]
-        
-        mock_root.children = [mock_func_node, mock_class_node]
-        mock_tree.root_node = mock_root
-        mock_parser.parse.return_value = mock_tree
-        
-        chunker.parsers = {"python": mock_parser}
-        
-        content = "def test_func():\n    pass\n\nclass TestClass:\n    pass"
-        
-        with patch.object(chunker, '_extract_code_units') as mock_extract:
-            mock_extract.return_value = [
-                {
-                    "type": "function",
-                    "name": "test_func",
-                    "start_pos": 0,
-                    "end_pos": 25
-                },
-                {
-                    "type": "class", 
-                    "name": "TestClass",
-                    "start_pos": 27,
-                    "end_pos": 50
-                }
-            ]
-            
-            chunks = chunker._ast_based_chunking(content, "python")
-            assert len(chunks) >= 2
-            assert any(chunk.chunk_type == "code" for chunk in chunks)
-    
-    def test_ast_chunking_with_no_code_units(self):
-        """Test AST chunking when no significant code structures found."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        mock_parser = Mock()
-        mock_tree = Mock()
-        mock_root = Mock()
-        mock_root.children = []
-        mock_tree.root_node = mock_root
-        mock_parser.parse.return_value = mock_tree
-        
-        chunker.parsers = {"python": mock_parser}
-        
-        content = "# Just a comment\nprint('hello')"
-        
-        with patch.object(chunker, '_extract_code_units', return_value=[]), \
-             patch.object(chunker, '_enhanced_chunking') as mock_enhanced:
-            mock_enhanced.return_value = [
-                Chunk(content=content, start_pos=0, end_pos=len(content), chunk_index=0)
-            ]
-            
-            chunks = chunker._ast_based_chunking(content, "python")
-            mock_enhanced.assert_called_once_with(content, "python")
-
-
-class TestCodeUnitExtraction:
-    """Test extraction of code units from AST."""
-    
-    def test_extract_python_function(self):
-        """Test extraction of Python function from AST."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock function node
-        mock_func_node = Mock()
-        mock_func_node.type = "function_definition"
-        mock_func_node.start_byte = 0
-        mock_func_node.end_byte = 30
-        mock_name_node = Mock()
-        mock_name_node.type = "identifier"
-        mock_name_node.start_byte = 4
-        mock_name_node.end_byte = 13
-        mock_func_node.children = [mock_name_node]
-        
-        content = "def test_func():\n    pass"
-        code_units = []
-        
-        chunker._traverse_python(mock_func_node, content, code_units)
-        
-        assert len(code_units) == 1
-        assert code_units[0]["type"] == "function"
-        assert code_units[0]["name"] == "test_func"
-        assert code_units[0]["start_pos"] == 0
-        assert code_units[0]["end_pos"] == 30
-    
-    def test_extract_python_class(self):
-        """Test extraction of Python class from AST."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock class node
-        mock_class_node = Mock()
-        mock_class_node.type = "class_definition"
-        mock_class_node.start_byte = 0
-        mock_class_node.end_byte = 40
-        mock_name_node = Mock()
-        mock_name_node.type = "identifier"
-        mock_name_node.start_byte = 6
-        mock_name_node.end_byte = 15
-        mock_class_node.children = [mock_name_node]
-        
-        content = "class TestClass:\n    pass"
-        code_units = []
-        
-        chunker._traverse_python(mock_class_node, content, code_units)
-        
-        assert len(code_units) == 1
-        assert code_units[0]["type"] == "class"
-        assert code_units[0]["name"] == "TestClass"
-        assert code_units[0]["start_pos"] == 0
-        assert code_units[0]["end_pos"] == 40
-    
-    def test_extract_javascript_function(self):
-        """Test extraction of JavaScript function from AST.""" 
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock function declaration node
-        mock_func_node = Mock()
-        mock_func_node.type = "function_declaration"
-        mock_func_node.start_byte = 0
-        mock_func_node.end_byte = 35
-        mock_name_node = Mock()
-        mock_name_node.type = "identifier"
-        mock_name_node.start_byte = 9
-        mock_name_node.end_byte = 17
-        mock_func_node.children = [mock_name_node]
-        
-        content = "function testFunc() { return 42; }"
-        code_units = []
-        
-        chunker._traverse_js_ts(mock_func_node, content, code_units, "javascript")
-        
-        assert len(code_units) == 1
-        assert code_units[0]["type"] == "function"
-        assert code_units[0]["name"] == "testFunc"
-        assert code_units[0]["start_pos"] == 0
-        assert code_units[0]["end_pos"] == 35
-    
-    def test_extract_javascript_class(self):
-        """Test extraction of JavaScript class from AST."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock class declaration node
-        mock_class_node = Mock()
-        mock_class_node.type = "class_declaration"
-        mock_class_node.start_byte = 0
-        mock_class_node.end_byte = 50
-        mock_name_node = Mock()
-        mock_name_node.type = "identifier"
-        mock_name_node.start_byte = 6
-        mock_name_node.end_byte = 15
-        mock_class_node.children = [mock_name_node]
-        
-        content = "class TestClass { constructor() {} }"
-        code_units = []
-        
-        chunker._traverse_js_ts(mock_class_node, content, code_units, "javascript")
-        
-        assert len(code_units) == 1
-        assert code_units[0]["type"] == "class"
-        assert code_units[0]["name"] == "TestClass"
-        assert code_units[0]["start_pos"] == 0
-        assert code_units[0]["end_pos"] == 50
-
-
-class TestLargeCodeUnitSplitting:
-    """Test splitting of large code units."""
-    
-    def test_split_large_function_with_ast(self):
-        """Test splitting large function using AST-based approach."""
-        config = ChunkingConfig(chunk_size=800, max_chunk_size=1600, max_function_chunk_size=3200)
-        chunker = EnhancedChunker(config)
-        
-        # Mock parser for recursive AST splitting
-        mock_parser = Mock()
-        mock_tree = Mock()
-        mock_root = Mock()
-        mock_body = Mock()
-        mock_body.type = "block"
-        mock_body.start_byte = 20
-        mock_body.end_byte = 200
-        mock_body.children = [
-            Mock(type="if_statement", start_byte=25, end_byte=100),
-            Mock(type="for_statement", start_byte=105, end_byte=180)
-        ]
-        mock_root.children = [mock_body]
-        mock_tree.root_node = mock_root
-        mock_parser.parse.return_value = mock_tree
-        
-        chunker.parsers = {"python": mock_parser}
-        
-        large_content = "def large_func():\n" + "    # code\n" * 20
-        
-        with patch.object(chunker, '_extract_function_blocks') as mock_extract:
-            mock_extract.return_value = [
-                {"type": "statement_sequence", "start_pos": 0, "end_pos": 100},
-                {"type": "statement_sequence", "start_pos": 100, "end_pos": 200}
-            ]
-            
-            chunks = chunker._split_large_code_unit(
-                large_content, 0, "function", "python"
-            )
-            
-            assert len(chunks) >= 1  # Relax assertion since the logic may create 1 chunk
-            assert all(chunk.chunk_type == "code" for chunk in chunks)
-    
-    def test_split_large_class_by_methods(self):
-        """Test splitting large class by methods."""
-        config = ChunkingConfig(chunk_size=800, max_chunk_size=1600, max_function_chunk_size=3200)
-        chunker = EnhancedChunker(config)
-        
-        # Mock parser and method extraction
-        mock_parser = Mock()
-        mock_tree = Mock()
-        mock_root = Mock()
-        mock_tree.root_node = mock_root
-        mock_parser.parse.return_value = mock_tree
-        
-        chunker.parsers = {"python": mock_parser}
-        
-        large_class = """class LargeClass:
-    def method1(self):
-        # lots of code
-        pass
-    
-    def method2(self):
-        # more code
-        pass"""
-        
-        with patch.object(chunker, '_extract_class_methods') as mock_extract:
-            mock_extract.return_value = [
-                {"name": "method1", "start_pos": 20, "end_pos": 80, "type": "method"},
-                {"name": "method2", "start_pos": 85, "end_pos": 140, "type": "method"}
-            ]
-            
-            chunks = chunker._split_large_code_unit(
-                large_class, 0, "class", "python" 
-            )
-            
-            assert len(chunks) >= 2
-            # First chunk should be class header
-            assert "class LargeClass:" in chunks[0].content or "method1" in chunks[0].content
-    
-    def test_split_large_unit_fallback_to_lines(self):
-        """Test fallback to line-based splitting when AST fails."""
-        config = ChunkingConfig(chunk_size=800, chunk_overlap=100, max_chunk_size=1600, max_function_chunk_size=3200)
-        chunker = EnhancedChunker(config)
-        
-        large_content = "def func():\n" + "    line\n" * 20
-        
-        with patch.object(chunker, '_chunk_large_code_block') as mock_fallback:
-            mock_fallback.return_value = [
-                Chunk(content="first part", start_pos=0, end_pos=50, chunk_index=0, chunk_type="code"),
-                Chunk(content="second part", start_pos=50, end_pos=100, chunk_index=1, chunk_type="code")
-            ]
-            
-            chunks = chunker._split_large_code_unit(
-                large_content, 0, "function", "unknown_language"
-            )
-            
-            mock_fallback.assert_called_once()
-            assert len(chunks) == 2
-
-
-class TestClassMethodExtraction:
-    """Test extraction of methods from classes."""
-    
-    def test_extract_python_class_methods(self):
-        """Test extraction of Python methods from class."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        content = "class TestClass:\n    def method1(self):\n        pass\n    \n    async def method2(self):\n        pass"
-        
-        # Calculate actual byte positions
-        method1_start = content.find("method1")
-        method1_end = method1_start + len("method1")
-        method2_start = content.find("method2")
-        method2_end = method2_start + len("method2")
-        
-        # Mock class node with method children
-        mock_class = Mock()
-        mock_method1 = Mock()
-        mock_method1.type = "function_definition"
-        mock_method1.start_byte = 20
-        mock_method1.end_byte = 60
-        mock_method1_name = Mock()
-        mock_method1_name.type = "identifier"
-        mock_method1_name.start_byte = method1_start
-        mock_method1_name.end_byte = method1_end
-        mock_method1_name.children = []
-        mock_method1.children = [mock_method1_name]
-        
-        mock_method2 = Mock()
-        mock_method2.type = "async_function_definition"
-        mock_method2.start_byte = 65
-        mock_method2.end_byte = 100
-        mock_method2_name = Mock()
-        mock_method2_name.type = "identifier"
-        mock_method2_name.start_byte = method2_start
-        mock_method2_name.end_byte = method2_end
-        mock_method2_name.children = []
-        mock_method2.children = [mock_method2_name]
-        
-        mock_class.children = [mock_method1, mock_method2]
-        
-        methods = chunker._extract_class_methods(mock_class, content, "python")
-        
-        assert len(methods) == 2
-        assert methods[0]["name"] == "method1"
-        assert methods[0]["type"] == "method"
-        assert methods[1]["name"] == "method2"
-        assert methods[1]["type"] == "method"
-    
-    def test_extract_javascript_class_methods(self):
-        """Test extraction of JavaScript methods from class."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock class node with method definition
-        mock_class = Mock()
-        mock_method = Mock()
-        mock_method.type = "method_definition"
-        mock_method.start_byte = 15
-        mock_method.end_byte = 50
-        mock_method.children = []  # Make sure children is iterable
-        mock_class.children = [mock_method]
-        
-        content = "class Test { method() { return 42; } }"
-        
-        with patch.object(chunker, '_get_js_method_name', return_value="method"):
-            methods = chunker._extract_class_methods(mock_class, content, "javascript")
-            
-            assert len(methods) == 1
-            assert methods[0]["name"] == "method"
-            assert methods[0]["type"] == "method"
-
-
-class TestFunctionBlockExtraction:
-    """Test extraction of blocks from functions."""
-    
-    def test_extract_function_blocks_small_function(self):
-        """Test that small functions are not split."""
-        config = ChunkingConfig(max_chunk_size=1600, max_function_chunk_size=3200)
-        chunker = EnhancedChunker(config)
-        
-        # Mock small function node
-        mock_func = Mock()
-        mock_func.start_byte = 0
-        mock_func.end_byte = 100  # Smaller than max_function_chunk_size
-        
-        content = "def small_func():\n    return 42"
-        
-        blocks = chunker._extract_function_blocks(mock_func, content, "python")
-        
-        assert len(blocks) == 1
-        assert blocks[0]["type"] == "whole_function"
-        assert blocks[0]["start_pos"] == 0
-        assert blocks[0]["end_pos"] == 100
-    
-    def test_extract_function_blocks_with_body(self):
-        """Test extraction of blocks from function with body."""
-        config = ChunkingConfig(chunk_size=800, chunk_overlap=100, max_chunk_size=1600, max_function_chunk_size=3200)
-        chunker = EnhancedChunker(config)
-        
-        # Mock large function with body
-        mock_func = Mock()
-        mock_func.start_byte = 0
-        mock_func.end_byte = 200  # Larger than max_function_chunk_size
-        
-        mock_body = Mock()
-        mock_body.type = "block"
-        mock_body.start_byte = 20
-        mock_body.end_byte = 180
-        
-        # Mock body children (statements)
-        mock_if = Mock()
-        mock_if.type = "if_statement"
-        mock_if.start_byte = 25
-        mock_if.end_byte = 80
-        
-        mock_for = Mock()
-        mock_for.type = "for_statement"
-        mock_for.start_byte = 85
-        mock_for.end_byte = 160
-        
-        mock_body.children = [mock_if, mock_for]
-        mock_func.children = [mock_body]
-        
-        content = "def large_func():\n    if True:\n        pass\n    for i in range(10):\n        pass"
-        
-        blocks = chunker._extract_function_blocks(mock_func, content, "python")
-        
-        # Should split into multiple blocks
-        assert len(blocks) >= 1
-        if len(blocks) > 1:
-            # First block should start at function beginning
-            assert blocks[0]["start_pos"] == 0
-    
-    def test_extract_function_signature(self):
-        """Test extraction of function signature."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock Python function node
-        mock_func = Mock()
-        mock_func.start_byte = 0
-        mock_colon = Mock()
-        mock_colon.type = ":"
-        mock_colon.start_byte = 25
-        mock_func.children = [mock_colon]
-        
-        content = "def example_func(a, b):\n    return a + b"
-        
-        signature = chunker._extract_function_signature(mock_func, content, "python")
-        assert signature == "def example_func(a, b):"
-        
-        # Mock JavaScript function node
-        mock_js_func = Mock()
-        mock_js_func.start_byte = 0
-        mock_body = Mock()
-        mock_body.type = "statement_block"
-        mock_body.start_byte = 20
-        mock_js_func.children = [mock_body]
-        
-        js_content = "function test() { return 42; }"
-        
-        signature = chunker._extract_function_signature(mock_js_func, js_content, "javascript")
-        assert signature == "function test() { re"  # Adjusted expectation based on byte range
-    
-    def test_get_js_method_name(self):
-        """Test extraction of JavaScript method names."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock method node with identifier
-        mock_method = Mock()
-        mock_name = Mock()
-        mock_name.type = "property_identifier"
-        mock_name.start_byte = 0
-        mock_name.end_byte = 6
-        mock_method.children = [mock_name]
-        
-        content = "method() { return 42; }"
-        
-        name = chunker._get_js_method_name(mock_method, content)
-        assert name == "method"
-        
-        # Test with no identifier
-        mock_empty = Mock()
-        mock_empty.children = []
-        
-        name = chunker._get_js_method_name(mock_empty, content)
-        assert name == ""
-
-
-class TestEdgeCasesAndErrorHandling:
-    """Test edge cases and error handling in AST chunking."""
-    
-    def test_ast_chunking_parser_exception(self):
-        """Test handling of parser exceptions during AST chunking."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        mock_parser = Mock()
-        mock_parser.parse.side_effect = Exception("Parser error")
-        chunker.parsers = {"python": mock_parser}
-        
-        content = "def test(): pass"
-        
-        with patch.object(chunker, '_enhanced_chunking') as mock_enhanced:
-            mock_enhanced.return_value = [
-                Chunk(content=content, start_pos=0, end_pos=len(content), chunk_index=0)
-            ]
-            
-            chunks = chunker._ast_based_chunking(content, "python")
-            mock_enhanced.assert_called_once_with(content, "python")
-    
-    def test_chunk_content_with_ast_strategy_fallback(self):
-        """Test main chunk_content method with AST strategy fallback."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        chunker.parsers = {}  # No parsers available
-        
-        content = "def test(): pass"
-        
-        with patch.object(chunker, '_enhanced_chunking') as mock_enhanced:
-            mock_enhanced.return_value = [
-                Chunk(content=content, start_pos=0, end_pos=len(content), chunk_index=0, chunk_type="text")
-            ]
-            
-            chunks = chunker.chunk_content(content, language="python")
-            
-            assert len(chunks) >= 1
-            assert isinstance(chunks[0], dict)  # Should be formatted
-            # Mock enhanced may not be called directly due to fallback paths
-    
-    def test_code_unit_extraction_empty_traversal(self):
-        """Test code unit extraction with empty AST traversal."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Mock empty node
-        mock_node = Mock()
-        mock_node.children = []
-        
-        content = "# just a comment"
-        
-        code_units = chunker._extract_code_units(mock_node, content, "python")
-        assert code_units == []
-
-
-class TestPerformanceOptimizations:
-    """Test performance-related functionality in chunking."""
-    
-    def test_chunking_performance_large_content(self):
-        """Test chunking performance with large content."""
-        config = ChunkingConfig(chunk_size=1000, strategy=ChunkingStrategy.ENHANCED)
-        chunker = EnhancedChunker(config)
-        
-        # Generate large content
-        large_content = "# Large file\n" + ("print('line')\n" * 1000)
-        
-        chunks = chunker._enhanced_chunking(large_content)
-        
-        # Should create multiple chunks efficiently
-        assert len(chunks) > 1
-        assert all(isinstance(chunk, Chunk) for chunk in chunks)
-        
-        # Check that chunks don't exceed size limits significantly
-        for chunk in chunks:
-            assert len(chunk.content) <= config.chunk_size * 1.2  # Allow some margin
-    
-    def test_boundary_detection_performance(self):
-        """Test boundary detection with various patterns."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Content with multiple boundary types
-        content = """
-        First paragraph with sentence. Another sentence!
-        
-        ## Header
-        
-        def function():
-            pass
-        
-        - List item
-        - Another item
-        
-        Final paragraph?
-        """
-        
-        boundary = chunker._find_enhanced_boundary(content, 0, 100)
-        
-        # Should find a reasonable boundary
-        assert 0 <= boundary <= len(content)
-        assert boundary != 100  # Should have found a better boundary
