--- tests/unit/test_chunking_focused_coverage.py
@@ -1,306 +1,0 @@
-"""Focused tests to cover specific missing lines in chunking.py.
-
-This test suite targets specific uncovered lines to boost coverage from 77% to 90%+.
-"""
-
-import pytest
-from unittest.mock import Mock, patch, MagicMock
-from src.chunking import EnhancedChunker
-from src.config.models import ChunkingConfig
-from src.config.enums import ChunkingStrategy
-
-
-class TestLanguageParserWarnings:
-    """Test lines 164-168: parser warning messages."""
-    
-    def test_unavailable_parser_warning(self):
-        """Test warning when parser is not available."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        
-        # Mock the language parsers to make one unavailable
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            chunker = EnhancedChunker(config)
-            
-            # Mock a language parser as unavailable
-            mock_logger = Mock()
-            chunker.logger = mock_logger
-            
-            # Mock language_parsers to have an unavailable parser
-            original_parsers = chunker.language_parsers.copy()
-            chunker.language_parsers["test_lang"] = (None, False)
-            
-            # Try to initialize parsers for unavailable language
-            languages = ["test_lang"]
-            chunker._initialize_parsers(languages)
-            
-            # Should have logged a warning
-            mock_logger.warning.assert_called()
-            call_args = mock_logger.warning.call_args[0][0]
-            assert "Parser for 'test_lang' is not available" in call_args
-            assert "pip install tree-sitter-test_lang" in call_args
-            
-            # Restore original parsers
-            chunker.language_parsers = original_parsers
-
-
-class TestPreContentChunking:
-    """Test lines 696-702: handling content before code units."""
-    
-    def test_pre_content_chunking_in_ast_strategy(self):
-        """Test chunking of content that appears before code units."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST, chunk_size=200)
-        chunker = EnhancedChunker(config)
-        
-        # Create content with text before code
-        content = """
-This is some explanatory text that appears before the code.
-It should be chunked separately from the code units.
-
-def example_function():
-    return "Hello World"
-"""
-        
-        # Mock code units that start after the text
-        mock_code_units = [
-            {
-                "type": "function",
-                "start_pos": content.find("def example_function"),
-                "end_pos": len(content),
-                "content": "def example_function():\n    return \"Hello World\""
-            }
-        ]
-        
-        # Mock _extract_code_units to return our mock units
-        with patch.object(chunker, '_extract_code_units', return_value=mock_code_units):
-            # Mock _chunk_text_content to verify it gets called
-            with patch.object(chunker, '_chunk_text_content') as mock_chunk_text:
-                mock_chunk_text.return_value = [
-                    {
-                        "content": "This is some explanatory text that appears before the code.",
-                        "start_pos": 0,
-                        "end_pos": 100
-                    }
-                ]
-                
-                chunks = chunker._chunk_with_ast(content, "python")
-                
-                # Should have called _chunk_text_content for pre-content
-                mock_chunk_text.assert_called()
-
-
-class TestErrorHandlingPaths:
-    """Test various error handling code paths."""
-    
-    def test_ast_parsing_exception_handling(self):
-        """Test handling of exceptions during AST parsing."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        content = "def test(): pass"
-        
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            with patch('src.chunking.PYTHON_AVAILABLE', True):
-                # Mock parser to raise exception
-                mock_parser = Mock()
-                mock_parser.parse.side_effect = Exception("Parser failed")
-                
-                with patch('src.chunking.Parser', return_value=mock_parser):
-                    with patch('src.chunking.tspython') as mock_tspython:
-                        mock_tspython.language.return_value = Mock()
-                        
-                        # Should fallback to enhanced chunking
-                        chunks = chunker._chunk_with_ast(content, "python")
-                        
-                        # Should still return chunks (from fallback)
-                        assert isinstance(chunks, list)
-    
-    def test_language_detection_edge_cases(self):
-        """Test edge cases in language detection."""
-        config = ChunkingConfig()
-        chunker = EnhancedChunker(config)
-        
-        # Test with URL that has no clear language
-        url_no_ext = "http://example.com/somefile"
-        language = chunker._detect_language(url_no_ext, "Some content")
-        assert language in ["text", "unknown", "markdown"]  # Reasonable fallbacks
-        
-        # Test with content that has no code fences
-        content_no_fences = "This is plain text with no code blocks."
-        language = chunker._detect_language("", content_no_fences)
-        assert language in ["text", "unknown", "markdown"]
-
-
-class TestChunkingBoundariesAndLimits:
-    """Test chunking boundaries and size limits."""
-    
-    def test_chunk_size_validation_edge_cases(self):
-        """Test edge cases in chunk size validation."""
-        config = ChunkingConfig(chunk_size=50, chunk_overlap=10)
-        chunker = EnhancedChunker(config)
-        
-        # Test with content exactly at chunk size limit
-        content = "x" * 50  # Exactly chunk_size
-        chunks = chunker.chunk_content(content, "test.txt", "http://test.txt")
-        
-        assert len(chunks) >= 1
-        assert all(len(chunk["content"]) <= config.chunk_size + config.chunk_overlap for chunk in chunks)
-    
-    def test_very_small_chunk_sizes(self):
-        """Test handling of very small chunk sizes."""
-        config = ChunkingConfig(chunk_size=20, chunk_overlap=5)
-        chunker = EnhancedChunker(config)
-        
-        content = "This is a test content that should be split into many small chunks."
-        chunks = chunker.chunk_content(content, "test.txt", "http://test.txt")
-        
-        assert len(chunks) > 3  # Should create multiple small chunks
-        assert all(chunk["content"] for chunk in chunks)  # All chunks should have content
-
-
-class TestRecursiveSplittingSimplified:
-    """Test recursive splitting with proper configuration."""
-    
-    def test_recursive_method_splitting_simplified(self):
-        """Test recursive splitting of large methods."""
-        config = ChunkingConfig(
-            strategy=ChunkingStrategy.AST,
-            chunk_size=500,
-            chunk_overlap=50,
-            max_function_chunk_size=200  # Smaller than chunk_size to trigger splitting
-        )
-        chunker = EnhancedChunker(config)
-        
-        # Create large method content
-        large_method_content = """
-def very_large_method(self):
-    # This method has many lines to exceed max_function_chunk_size
-    step1 = "first processing step"
-    step2 = "second processing step"
-    step3 = "third processing step"
-    step4 = "fourth processing step"
-    step5 = "fifth processing step"
-    step6 = "sixth processing step"
-    
-    for i in range(100):
-        if i % 2 == 0:
-            print(f"Even: {i}")
-        else:
-            print(f"Odd: {i}")
-    
-    return "completed"
-"""
-        
-        # Mock _split_large_code_unit to be called for large content
-        original_split = chunker._split_large_code_unit
-        
-        def mock_split_large_unit(content, start_pos, unit_type, language):
-            # Return multiple chunks to simulate splitting
-            mid_point = len(content) // 2
-            from src.chunking import Chunk
-            chunk1 = Chunk(
-                content=content[:mid_point],
-                start_pos=start_pos,
-                end_pos=start_pos + mid_point,
-                metadata={"parent_class": True}
-            )
-            chunk2 = Chunk(
-                content=content[mid_point:],
-                start_pos=start_pos + mid_point,
-                end_pos=start_pos + len(content),
-                metadata={"parent_class": True}
-            )
-            return [chunk1, chunk2]
-        
-        with patch.object(chunker, '_split_large_code_unit', side_effect=mock_split_large_unit):
-            chunks = chunker.chunk_content(large_method_content, "test.py", "http://test.py")
-            
-            # Should have created chunks
-            assert len(chunks) > 0
-
-
-class TestTreeSitterNodeIteration:
-    """Test tree-sitter node iteration and traversal."""
-    
-    def test_node_children_iteration(self):
-        """Test iteration over tree-sitter node children."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        content = "def test(): return True"
-        
-        with patch('src.chunking.TREE_SITTER_AVAILABLE', True):
-            with patch('src.chunking.PYTHON_AVAILABLE', True):
-                # Create mock tree with iterable children
-                mock_tree = Mock()
-                mock_root = Mock()
-                
-                # Create mock function node
-                mock_func_node = Mock()
-                mock_func_node.type = "function_definition"
-                mock_func_node.start_byte = 0
-                mock_func_node.end_byte = len(content)
-                mock_func_node.children = []  # Empty children to avoid recursion
-                
-                mock_root.children = [mock_func_node]
-                mock_tree.root_node = mock_root
-                
-                mock_parser = Mock()
-                mock_parser.parse.return_value = mock_tree
-                
-                with patch('src.chunking.Parser', return_value=mock_parser):
-                    with patch('src.chunking.tspython') as mock_tspython:
-                        mock_tspython.language.return_value = Mock()
-                        
-                        # Test node iteration during code unit extraction
-                        code_units = chunker._extract_code_units(mock_tree, "python", 1000)
-                        
-                        # Should have found the function
-                        assert isinstance(code_units, list)
-
-
-class TestChunkMetadataHandling:
-    """Test chunk metadata handling and preservation."""
-    
-    def test_chunk_metadata_preservation(self):
-        """Test that chunk metadata is properly preserved."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.ENHANCED, chunk_size=100)
-        chunker = EnhancedChunker(config)
-        
-        content = """
-# This is a header comment
-def function_with_metadata():
-    \"\"\"A function with documentation.\"\"\"
-    return "result"
-"""
-        
-        chunks = chunker.chunk_content(content, "test.py", "http://test.py")
-        
-        # Verify chunks have basic metadata
-        assert len(chunks) > 0
-        for chunk in chunks:
-            assert "content" in chunk
-            assert "start_pos" in chunk
-            assert "end_pos" in chunk
-            # May have additional metadata depending on strategy
-    
-    def test_function_signature_extraction(self):
-        """Test function signature extraction from code."""
-        config = ChunkingConfig(strategy=ChunkingStrategy.AST)
-        chunker = EnhancedChunker(config)
-        
-        # Test Python function signature
-        python_content = "def example_function(param1: str, param2: int = 5) -> str:"
-        
-        # Mock function signature extraction
-        if hasattr(chunker, '_extract_function_signature'):
-            mock_node = Mock()
-            mock_node.start_byte = 0
-            mock_node.end_byte = len(python_content)
-            
-            try:
-                signature = chunker._extract_function_signature(mock_node, python_content, "python")
-                assert isinstance(signature, str)
-                assert "example_function" in signature
-            except (AttributeError, TypeError):
-                # Method might not exist or have different signature
-                pass
